Fix planning document title
