+++ src/components/thumbnail/tab-panel-documents-subsection-panel.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import { ThumbnailDocumentItem } from "./thumbnail-document-item";
 import { DocumentModelType } from "../../models/document/document";
-import {
-  isPlanningType, isProblemType, isPublishedType, SupportPublication
-} from "../../models/document/document-types";
+import { isPublishedType, SupportPublication } from "../../models/document/document-types";
+import { getDocumentDisplayTitle } from "../../models/document/document-utils";
 import { IStores } from "../../models/stores/stores";
 import { NavTabSectionModelType } from "../../models/view/nav-tabs";
 
@@ -29,11 +28,7 @@ function getDocumentCaption(stores: IStores, document: DocumentModelType) {
   const user = _class?.getUserById(uid);
   const userName = user?.displayName;
   const namePrefix = isPublishedType(type) ? `${userName}: ` : "";
-  const title = isProblemType(type)
-                  ? problem.title
-                  : isPlanningType(type)
-                      ? `${problem.title}: Planning`
-                      : document.getDisplayTitle(appConfig);
+  const title = getDocumentDisplayTitle(document, appConfig, problem);
   return `${namePrefix}${title}`;
 }
